Scope util::URL
