+++ libmamba/src/core/validate.cpp
@@ -15,9 +15,9 @@
 
 #include "mamba/core/fetch.hpp"
 #include "mamba/core/output.hpp"
-#include "mamba/core/url.hpp"
 #include "mamba/core/validate.hpp"
 #include "mamba/util/string.hpp"
+#include "mamba/util/url.hpp"
 
 namespace mamba
 {
@@ -1441,7 +1441,7 @@ namespace mamba::validation
             auto tmp_dir = std::make_unique<mamba::TemporaryDirectory>();
             auto tmp_metadata_path = tmp_dir->path() / "key_mgr.json";
 
-            const auto url = mamba::URL::parse(base_url + "/key_mgr.json");
+            const auto url = mamba::util::URL::parse(base_url + "/key_mgr.json");
 
             auto dl_target = std::make_unique<mamba::DownloadTarget>(
                 "key_mgr.json",
@@ -1604,7 +1604,7 @@ namespace mamba::validation
             auto tmp_dir = std::make_unique<mamba::TemporaryDirectory>();
             auto tmp_metadata_path = tmp_dir->path() / "pkg_mgr.json";
 
-            const auto url = mamba::URL::parse(base_url + "/pkg_mgr.json");
+            const auto url = mamba::util::URL::parse(base_url + "/pkg_mgr.json");
 
             auto dl_target = std::make_unique<mamba::DownloadTarget>(
                 "pkg_mgr.json",
